test(events): 拆分 EventBusTests 事件测试辅助类型
- 拆分 EventBusTestsEvent 到独立文件以消除 EventBusTests.cs 的 MA0048 警告

- 补充提取事件类型的 XML 文档并保持 EventBusTests 行为不变
